+++ R/function_Spectra_qualityMetrics.R
@@ -44,7 +44,7 @@
 qualityMetrics <- function(object) {
     if (is(object, "Spectra"))
         .metrics <- c(
-            "rtDuration", "rtOverTicQuantile", "rtOverMsQuarters",
+            "rtDuration", "rtOverTicQuantiles", "rtOverMsQuarters",
             "ticQuartileToQuartileLogRatio", "numberSpectra", 
             "medianPrecursorMz", "rtIqr", "rtIqrRate", "areaUnderTic", 
             "areaUnderTicRtQuantiles", "extentIdentifiedPrecursorIntensity",
@@ -58,7 +58,7 @@ qualityMetrics <- function(object) {
     
     if (is(object, "MsExperiment"))
         .metrics <-  c(
-            "rtDuration", "rtOverTicQuantile", "rtOverMsQuarters",
+            "rtDuration", "rtOverTicQuantiles", "rtOverMsQuarters",
             "ticQuartileToQuartileLogRatio", "numberSpectra", 
             "medianPrecursorMz", "rtIqr", "rtIqrRate", "areaUnderTic", 
             "areaUnderTicRtQuantiles", "extentIdentifiedPrecursorIntensity",